+++ src/box/sensorLayouts/sensorDefinitions/pms3003_pm10.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  title: 'PM10',
+  unit: 'µg/m³',
+  sensorType: 'PMS 3003',
+  icon: 'osem-cloud'
+};